authorize bug fixed

+++ Restaurants.Application/Dishs/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,25 +1,19 @@
-﻿using AutoMapper;
-using MediatR;
+﻿using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurant.Domain.Exceptions;
+using Restaurant.Domain.Interfaces;
 using Restaurant.Domain.Repositories;
 
 namespace Restaurants.Application.Dishs.Commands.DeleteDishes
 {
-    public class DeleteDishesForRestaurantCommandHandler : IRequestHandler<DeleteDishesForRestaurantCommand>
+    public class DeleteDishesForRestaurantCommandHandler(
+        IDishRepository _dishRepository,
+        IRestaurantRepository _restaurantRepository,
+        ILogger<DeleteDishesForRestaurantCommandHandler> _logger,
+        IRestaurantAuthorizationService _restaurantAuthorizationService
+        ) : IRequestHandler<DeleteDishesForRestaurantCommand>
     {
-        private readonly IDishRepository _dishRepository;
-        private readonly IRestaurantRepository _restaurantRepository;
-        private readonly IMapper _mapper;
-        private readonly ILogger<DeleteDishesForRestaurantCommandHandler> _logger;
-
-        public DeleteDishesForRestaurantCommandHandler(IDishRepository dishRepository, IRestaurantRepository restaurantRepository, IMapper mapper, ILogger<DeleteDishesForRestaurantCommandHandler> logger)
-        {
-            _dishRepository = dishRepository;
-            _restaurantRepository = restaurantRepository;
-            _mapper = mapper;
-            _logger = logger;
-        }
+    
 
         public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -29,6 +23,11 @@ namespace Restaurants.Application.Dishs.Commands.DeleteDishes
             {
                 _logger.LogWarning("Restaurant with ID {RestaurantId} not found", request.RestaurantId);
                 throw new NotFoundException($"Restaurant with ID {request.RestaurantId} not found. to delete all its dishes");
+            }
+            if (!_restaurantAuthorizationService.Authorize(restaurant, RestaurantOperations.delete))
+            {
+                _logger.LogWarning("Unauthorized attempt to delete Dishes for restaurant with ID: {RestaurantId}", request.RestaurantId);
+                throw new ForbiddenExceptions("You do not have permission to delete this restaurant.");
             }
             await _dishRepository.DeleteAllAsync(restaurant.Dishes);
